More aggressive transaction retry in --slave-parallel-mode=aggressive
When transaction T1 in parallel replication needs to deadlock kill another
transaction T3, record the wait_for_commit and the sub_id of T1 in the
rpl_group_info of T3.

Then in aggressive mode, when we do the retry of T3, wait for T1 to complete
before starting the retry. In other modes, we still wait for all prior
transactions to complete before the retry (eg. T2 if we have T1 T2 T3).

This is a tentative patch to study ways to improve in-order parallel
replication of loads with many lock conflicts between transactions. This
patch might reduce the impact of transaction retries due to such conflicts,
as they allow the conflicting transactions to be retried earlier. The
downside is that such transactions may conflict again with other prior
transactions. At worst, up to (N-1) conflicts could occur, if
--slave-paralle-mode=N.